requirements.txt
FastAPI
SQLAlchemy
mysql-connector-python
pydantic
python-jose
uvicorn
alembic
python-dotenv
PyJWT